% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_file.R,
%   R/estimate_black_earth_parameters.R
\name{estimate_black_earth_parameters}
\alias{estimate_black_earth_parameters}
\alias{estimate_black_earth_parameters}
\title{Create parameter set for Black Earth Creek.}
\usage{
estimate_black_earth_parameters()

estimate_black_earth_parameters()
}
\arguments{
\item{filename}{Name of the PRMS 'par_name' file to munge.}
}
\value{
List containing the name, value, and description of each PRMS dimension.

None.
}
\description{
This function demonstrates how the other functions in this package may be used to
\enumerate{
\item read in dimensions and default parameters generated by using the PRMS \code{-print} directive,
\item modify the parameters using custom functions or lookup tables, and
\item write the resulting parameters to a new PRMS parameter file.
}

This function demonstrates how the other functions in this package may be used to
\enumerate{
\item read in dimensions and default parameters generated by using the PRMS \code{-print} directive,
\item modify the parameters using custom functions or lookup tables, and
\item write the resulting parameters to a new PRMS parameter file.
}
}
