% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prms_dimensions.R
\name{get_prms_dimensions}
\alias{get_prms_dimensions}
\title{Get PRMS dimension information from a PRMS 'par_name' file.}
\usage{
get_prms_dimensions(filename)
}
\arguments{
\item{filename}{Name of the PRMS 'par_name' file to munge.}
}
\value{
List containing the name, value, and description of each PRMS dimension.
}
\description{
The input file is the output from PRMS run with the \code{-print} option. This function will
not work with an ordinary PRMS parameter file.
}
