% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ggo.R
\name{read_ggo}
\alias{read_ggo}
\title{Read GSFLOW *.ggo file.}
\usage{
read_ggo(filename, start_date_str = "01/01/1970")
}
\arguments{
\item{filename}{Name of the GSFLOW *.ggo parameter file to munge.}

\item{start_date_str}{}
}
\value{
Calculate an observation date and return the observations stored in the 'ggo' file.
}
\description{
Read GSFLOW *.ggo file.
}
