% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prms_parameter_file.R
\name{read_prms_parameter_file}
\alias{read_prms_parameter_file}
\title{Read PRMS parameter from a PRMS model parameter file.}
\usage{
read_prms_parameter_file(filename)
}
\arguments{
\item{filename}{Name of the PRMS parameter file to munge.}
}
\value{
List containing the name, value, description, default value, etc. of each PRMS parameter.
}
\description{
The input file is \emph{NOT} the output from PRMS run with the \code{-print} option. This function will
\emph{ONLY} work with an ordinary PRMS parameter file.
}
