% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ssf.R
\name{read_ssf}
\alias{read_ssf}
\title{Read PEST/TSPROC *.ssf file.}
\usage{
read_ssf(filename)
}
\arguments{
\item{filename}{Name of the PEST/TSPROC *.ssf file to munge.}

\item{start_date_str}{}
}
\value{
Return the observations stored in the 'ssf' file.
}
\description{
Read PEST/TSPROC *.ssf file.
}
