% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statvar.R
\name{read_statvar}
\alias{read_statvar}
\title{Read model output from a 'statvar' file.}
\usage{
read_statvar(filename)
}
\arguments{
\item{filename}{Name of a PRMS model 'statvar' file.}
}
\value{
List containing variable names, unit numbers (HRU numbers etc.), and a dataframe
containing all statvar file output.
}
\description{
Read model output from a 'statvar' file.
}
