% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_prms_dimension.R
\name{write_prms_dimension}
\alias{write_prms_dimension}
\title{Write PRMS dimension information to a PRMS parameter file.}
\usage{
write_prms_dimension(fname = "", dimname = "one", dimlength = 1,
  dHeader = FALSE)
}
\arguments{
\item{filename}{Name of the PRMS parameter file to write to.}
}
\value{
None.
}
\description{
This function is designed to be called once for each dimension written to the file.
For example it will be called once for the 'nmonth' dimension, again for the 'ntemp'
dimension, etc.
}
