% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_prms_header.R
\name{write_prms_header}
\alias{write_prms_header}
\title{Write PRMS header to a PRMS parameter file.}
\usage{
write_prms_header(filename)
}
\arguments{
\item{filename}{Name of the PRMS parameter file to write to.}
}
\value{
None.
}
\description{
Write PRMS header to a PRMS parameter file.
}
