% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_prms_parameter.R
\name{write_prms_parameter}
\alias{write_prms_parameter}
\title{Write PRMS parameter information to a PRMS parameter file.}
\usage{
write_prms_parameter(filename = "", parname = "basin", dimname1 = "one",
  dimname2 = "", type = 2, values, dHeader = FALSE)
}
\arguments{
\item{filename}{Name of the PRMS parameter file to write to.}
}
\value{
None.
}
\description{
This function is designed to be called once for each parameter written to the file.
For example it will be called once for the 'smidx_exp' parameters, again for the 'den_init'
parameters, etc.
}
